spacer-backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
cloudinary==1.35.0
sendinblue==2.0.5.1
marshmallow==3.20.1
pytest==7.4.3
python-dateutil==2.8.2
flasgger==0.9.5
